requirements.txt
numpy
pandas
pdfminer.six
python-docx
nltk
spacy
scikit-learn
transformers
torch
sentence-transformers
flask
gunicorn
python-dotenv
tqdm
rich
